refactor: move int tests to integrationtests folder
