+++ src/integrationtests/java/com/aws/greengrass/integrationtests/ConfigTest.java
@@ -6,38 +6,64 @@
 package com.aws.greengrass.integrationtests;
 
 import com.aws.greengrass.clientdevices.auth.CertificateManager;
+import com.aws.greengrass.config.Topics;
+import com.aws.greengrass.config.UpdateBehaviorTree;
 import com.aws.greengrass.dependency.State;
 import com.aws.greengrass.lifecyclemanager.GlobalStateChangeListener;
 import com.aws.greengrass.lifecyclemanager.GreengrassService;
 import com.aws.greengrass.lifecyclemanager.Kernel;
+import com.aws.greengrass.mqtt.bridge.BridgeConfig;
 import com.aws.greengrass.mqtt.bridge.MQTTBridge;
 import com.aws.greengrass.mqtt.bridge.MQTTBridgeTest;
+import com.aws.greengrass.mqtt.bridge.TopicMapping;
 import com.aws.greengrass.testcommons.testutilities.GGExtension;
 import com.aws.greengrass.testcommons.testutilities.UniqueRootPathExtension;
+import com.aws.greengrass.util.Utils;
+import com.fasterxml.jackson.databind.exc.InvalidFormatException;
+import com.github.grantwest.eventually.EventuallyLambdaMatcher;
 import io.moquette.BrokerConstants;
 import io.moquette.broker.Server;
 import io.moquette.broker.config.IConfig;
 import io.moquette.broker.config.MemoryConfig;
+import org.eclipse.paho.client.mqttv3.MqttException;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.api.io.TempDir;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.io.IOException;
 import java.nio.file.Path;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Properties;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ScheduledThreadPoolExecutor;
+import java.util.concurrent.Semaphore;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.function.Supplier;
 
+import static com.aws.greengrass.componentmanager.KernelConfigResolver.CONFIGURATION_CONFIG_KEY;
+import static com.aws.greengrass.testcommons.testutilities.ExceptionLogProtector.ignoreExceptionOfType;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.is;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.mock;
 
 @ExtendWith({GGExtension.class, UniqueRootPathExtension.class, MockitoExtension.class})
 public class ConfigTest {
+    private static final long TEST_TIME_OUT_SEC = 30L;
+    private static final Supplier<UpdateBehaviorTree> MERGE_UPDATE_BEHAVIOR =
+            () -> new UpdateBehaviorTree(UpdateBehaviorTree.UpdateBehavior.MERGE, System.currentTimeMillis());
+
     @TempDir
     Path rootDir;
 
@@ -81,14 +107,208 @@ public class ConfigTest {
                 bridgeRunning.countDown();
             }
         };
-        kernel.getContext().addGlobalStateChangeListener(listener);
-        kernel.launch();
-        assertTrue(bridgeRunning.await(10, TimeUnit.SECONDS));
+        try {
+            kernel.getContext().addGlobalStateChangeListener(listener);
+            kernel.launch();
+            assertTrue(bridgeRunning.await(10, TimeUnit.SECONDS));
+        } finally {
+            kernel.getContext().removeGlobalStateChangeListener(listener);
+        }
     }
 
-    @Test // TODO add other tests
-    void GIVEN_bridge_WHEN_kernel_started_THEN_bridge_starts() throws Exception {
+    @Test
+    void GIVEN_Greengrass_with_mqtt_bridge_WHEN_multiple_config_changes_consecutively_THEN_bridge_reinstalls_once(ExtensionContext context)
+            throws Exception {
+        ignoreExceptionOfType(context, InterruptedException.class);
         startKernelWithConfig("config.yaml");
+
+        CountDownLatch bridgeRestarted = new CountDownLatch(1);
+        AtomicInteger numRestarts = new AtomicInteger();
+
+        kernel.getContext().addGlobalStateChangeListener((GreengrassService service, State was, State newState) -> {
+            if (service.getName().equals(MQTTBridge.SERVICE_NAME) && newState.equals(State.NEW)) {
+                numRestarts.incrementAndGet();
+                bridgeRestarted.countDown();
+            }
+        });
+
+        Topics config = kernel.locate(MQTTBridge.SERVICE_NAME).getConfig()
+                .lookupTopics(CONFIGURATION_CONFIG_KEY);
+
+        config.updateFromMap(Utils.immutableMap(BridgeConfig.KEY_CLIENT_ID, "new_client_id"), MERGE_UPDATE_BEHAVIOR.get());
+        config.updateFromMap(Utils.immutableMap(BridgeConfig.KEY_BROKER_URI, "tcp://newbroker:1234"), MERGE_UPDATE_BEHAVIOR.get());
+
+        assertTrue(bridgeRestarted.await(TEST_TIME_OUT_SEC, TimeUnit.SECONDS));
+        assertEquals(1, numRestarts.get());
     }
 
+    @Test
+    void GIVEN_Greengrass_with_mqtt_bridge_WHEN_multiple_serialized_config_changes_occur_THEN_bridge_reinstalls_multiple_times(ExtensionContext context)
+            throws Exception {
+        ignoreExceptionOfType(context, InterruptedException.class);
+        startKernelWithConfig("config.yaml");
+
+        Semaphore bridgeRestarted = new Semaphore(1);
+        bridgeRestarted.acquire();
+
+        kernel.getContext().addGlobalStateChangeListener((GreengrassService service, State was, State newState) -> {
+            if (service.getName().equals(MQTTBridge.SERVICE_NAME) && newState.equals(State.RUNNING)) {
+                bridgeRestarted.release();
+            }
+        });
+
+        Topics config = kernel.locate(MQTTBridge.SERVICE_NAME).getConfig()
+                .lookupTopics(CONFIGURATION_CONFIG_KEY);
+
+        int numRestarts = 5;
+        for (int i = 0; i < numRestarts; i++) {
+            // change the configuration and wait for bridge to restart
+            config.updateFromMap(Utils.immutableMap(BridgeConfig.KEY_BROKER_URI, String.format("tcp://brokeruri:%d", i)), MERGE_UPDATE_BEHAVIOR.get());
+            assertTrue(bridgeRestarted.tryAcquire(TEST_TIME_OUT_SEC, TimeUnit.SECONDS));
+        }
+    }
+
+    @Test
+    void GIVEN_Greengrass_with_mqtt_bridge_WHEN_clientId_config_changes_THEN_bridge_reinstalls() throws Exception {
+        startKernelWithConfig("config.yaml");
+
+        CountDownLatch bridgeRestarted = new CountDownLatch(1);
+        kernel.getContext().addGlobalStateChangeListener((GreengrassService service, State was, State newState) -> {
+            if (service.getName().equals(MQTTBridge.SERVICE_NAME) && newState.equals(State.NEW)) {
+                bridgeRestarted.countDown();
+            }
+        });
+
+        Topics config = kernel.locate(MQTTBridge.SERVICE_NAME).getConfig()
+                .lookupTopics(CONFIGURATION_CONFIG_KEY);
+        config.updateFromMap(Utils.immutableMap(BridgeConfig.KEY_CLIENT_ID, "new_client_id"), MERGE_UPDATE_BEHAVIOR.get());
+
+        assertTrue(bridgeRestarted.await(TEST_TIME_OUT_SEC, TimeUnit.SECONDS));
+    }
+
+    @Test
+    void GIVEN_Greengrass_with_mqtt_bridge_WHEN_valid_mqttTopicMapping_updated_THEN_mapping_updated(ExtensionContext context)
+            throws Exception {
+        ignoreExceptionOfType(context, MqttException.class);
+        startKernelWithConfig("config.yaml");
+        TopicMapping topicMapping = kernel.getContext().get(TopicMapping.class);
+        assertThat(topicMapping.getMapping().size(), is(equalTo(0)));
+
+        CountDownLatch bridgeRestarted = new CountDownLatch(1);
+        kernel.getContext().addGlobalStateChangeListener((GreengrassService service, State was, State newState) -> {
+            if (service.getName().equals(MQTTBridge.SERVICE_NAME) && newState.equals(State.NEW)) {
+                bridgeRestarted.countDown();
+            }
+        });
+
+        Topics mappingConfigTopics = kernel.locate(MQTTBridge.SERVICE_NAME).getConfig()
+                .lookupTopics(CONFIGURATION_CONFIG_KEY, BridgeConfig.KEY_MQTT_TOPIC_MAPPING);
+
+        mappingConfigTopics.replaceAndWait(Utils.immutableMap("m1",
+                Utils.immutableMap("topic", "mqtt/topic", "source", TopicMapping.TopicType.LocalMqtt.toString(),
+                        "target", TopicMapping.TopicType.IotCore.toString()), "m2",
+                Utils.immutableMap("topic", "mqtt/topic2", "source", TopicMapping.TopicType.LocalMqtt.toString(),
+                        "target", TopicMapping.TopicType.Pubsub.toString()), "m3",
+                Utils.immutableMap("topic", "mqtt/topic3", "source", TopicMapping.TopicType.LocalMqtt.toString(),
+                        "target", TopicMapping.TopicType.IotCore.toString())));
+
+        kernel.getContext().waitForPublishQueueToClear();
+        assertThat(topicMapping.getMapping().size(), is(equalTo(3)));
+        assertFalse(bridgeRestarted.await(2, TimeUnit.SECONDS));
+    }
+
+    @Test
+    void GIVEN_Greengrass_with_mqtt_bridge_WHEN_valid_mapping_provided_in_config_THEN_mapping_populated(ExtensionContext context)
+            throws Exception {
+        ignoreExceptionOfType(context, MqttException.class);
+        startKernelWithConfig("config_with_mapping.yaml");
+        TopicMapping topicMapping = kernel.getContext().get(TopicMapping.class);
+
+        assertThat(() -> topicMapping.getMapping().size(), EventuallyLambdaMatcher.eventuallyEval(is(5)));
+        Map<String, TopicMapping.MappingEntry> expectedMapping = new HashMap<>();
+        expectedMapping.put("mapping1",
+                new TopicMapping.MappingEntry("topic/to/map/from/local/to/cloud", TopicMapping.TopicType.LocalMqtt,
+                        TopicMapping.TopicType.IotCore));
+        expectedMapping.put("mapping2",
+                new TopicMapping.MappingEntry("topic/to/map/from/local/to/pubsub", TopicMapping.TopicType.LocalMqtt,
+                        TopicMapping.TopicType.Pubsub));
+        expectedMapping.put("mapping3",
+                new TopicMapping.MappingEntry("topic/to/map/from/local/to/cloud/2", TopicMapping.TopicType.LocalMqtt,
+                        TopicMapping.TopicType.IotCore));
+        expectedMapping.put("mapping4",
+                new TopicMapping.MappingEntry("topic/to/map/from/local/to/pubsub/2", TopicMapping.TopicType.LocalMqtt,
+                        TopicMapping.TopicType.Pubsub, "a-prefix"));
+        expectedMapping.put("mapping5",
+                new TopicMapping.MappingEntry("topic/to/map/from/local/to/cloud/3", TopicMapping.TopicType.LocalMqtt,
+                        TopicMapping.TopicType.IotCore, "a-prefix"));
+
+        assertEquals(expectedMapping, topicMapping.getMapping());
+    }
+
+    @Test
+    void GIVEN_Greengrass_with_mqtt_bridge_WHEN_empty_mqttTopicMapping_updated_THEN_mapping_not_updated(ExtensionContext context)
+            throws Exception {
+        ignoreExceptionOfType(context, MqttException.class);
+        startKernelWithConfig("config.yaml");
+        TopicMapping topicMapping = kernel.getContext().get(TopicMapping.class);
+        assertThat(topicMapping.getMapping().size(), is(equalTo(0)));
+
+        kernel.locate(MQTTBridge.SERVICE_NAME).getConfig()
+                .lookupTopics(CONFIGURATION_CONFIG_KEY, BridgeConfig.KEY_MQTT_TOPIC_MAPPING)
+                .replaceAndWait(Collections.emptyMap());
+        // Block until subscriber has finished updating
+        kernel.getContext().waitForPublishQueueToClear();
+        assertThat(topicMapping.getMapping().size(), is(equalTo(0)));
+    }
+
+    @Test
+    void GIVEN_Greengrass_with_mqtt_bridge_WHEN_mapping_updated_with_empty_THEN_mapping_removed(ExtensionContext context)
+            throws Exception {
+        ignoreExceptionOfType(context, MqttException.class);
+        startKernelWithConfig("config_with_mapping.yaml");
+        TopicMapping topicMapping = kernel.getContext().get(TopicMapping.class);
+
+        assertThat(() -> topicMapping.getMapping().size(), EventuallyLambdaMatcher.eventuallyEval(is(5)));
+        kernel.locate(MQTTBridge.SERVICE_NAME).getConfig()
+                .lookupTopics(CONFIGURATION_CONFIG_KEY, BridgeConfig.KEY_MQTT_TOPIC_MAPPING)
+                .replaceAndWait(Collections.emptyMap());
+        // Block until subscriber has finished updating
+        kernel.getContext().waitForPublishQueueToClear();
+        assertThat(topicMapping.getMapping().size(), is(equalTo(0)));
+    }
+
+    @Test
+    void GIVEN_Greengrass_with_mqtt_bridge_WHEN_invalid_mqttTopicMapping_updated_THEN_mapping_not_updated(ExtensionContext context)
+            throws Exception {
+        ignoreExceptionOfType(context, InvalidFormatException.class);
+        startKernelWithConfig("config.yaml");
+        TopicMapping topicMapping = kernel.getContext().get(TopicMapping.class);
+        assertThat(topicMapping.getMapping().size(), is(equalTo(0)));
+
+        CountDownLatch bridgeErrored = new CountDownLatch(1);
+        GlobalStateChangeListener listener = (GreengrassService service, State was, State newState) -> {
+            if (service.getName().equals(MQTTBridge.SERVICE_NAME) && service.getState().equals(State.ERRORED)) {
+                bridgeErrored.countDown();
+            }
+        };
+        kernel.getContext().addGlobalStateChangeListener(listener);
+
+        // Updating with invalid mapping (Providing type as Pubsub-Invalid)
+        Topics mappingConfigTopics = kernel.locate(MQTTBridge.SERVICE_NAME).getConfig()
+                .lookupTopics(CONFIGURATION_CONFIG_KEY, BridgeConfig.KEY_MQTT_TOPIC_MAPPING);
+
+        mappingConfigTopics.replaceAndWait(Utils.immutableMap("m1",
+                Utils.immutableMap("topic", "mqtt/topic", "source", TopicMapping.TopicType.LocalMqtt.toString(),
+                        "target", TopicMapping.TopicType.IotCore.toString()), "m2",
+                Utils.immutableMap("topic", "mqtt/topic2", "source", TopicMapping.TopicType.LocalMqtt.toString(),
+                        "target", "Pubsub-Invalid"), "m3",
+                Utils.immutableMap("topic", "mqtt/topic3", "source", TopicMapping.TopicType.LocalMqtt.toString(),
+                        "target", TopicMapping.TopicType.IotCore.toString())));
+
+        // Block until subscriber has finished updating
+        kernel.getContext().waitForPublishQueueToClear();
+        assertThat(topicMapping.getMapping().size(), is(equalTo(0)));
+
+        assertTrue(bridgeErrored.await(TEST_TIME_OUT_SEC, TimeUnit.SECONDS));
+    }
 }
